task/refactor (#27)
* chore: clear SonarCloud's open maintainability and security issues

Index the first element of a list the tests already hold as IReadOnlyList
rather than going through Enumerable.First (CA1826, 30 sites), and use a
collection expression for an empty array (CA1825).

IntegerConvertibleTests' theory rows held IConvertible instances, which xUnit
cannot serialize, so Test Explorer showed each theory as one row. The rows now
name the wrapper and each test builds the number from that name (xUnit1045).

Give the fonts.conf regex in LinuxSystemFontResolver a one-second match
timeout (S6444). A timeout lands in the catch that is already there, which logs
it and returns the directories read so far.

* refactor

* Descend the paragraph tree again when seeking its first and last leaf

The refactor turned the cast inside SeekFirstLeaf's and SeekLastLeaf's
loops into a pattern variable bound once, before the loop, and then
tested that variable in the loop's condition. The loop body moves obj
down the tree but never rebinds parEls, so the condition stayed true
for ever: every paragraph with any content spun there, adding an index
to the position list on each pass.

A test's Timeout does not stop that. Interruptibly runs the work on a
background thread, which keeps looping after the test has been failed,
so every paragraph test left one behind allocating without bound, and
the full suite ran the machine out of memory.

Testing obj on each pass, as the code did before the refactor, restores
the descent.

* chore: remove redundant casts reported by InspectCode

* chore: fix the InspectCode findings for array creation, unused variables and unused parameters

Implicitly type 70 array creations, drop or discard unused locals (keeping
the calls whose side effects matter), and remove unused parameters from
private and internal members along with their callers, or discard them in
lambdas. Public API is untouched.

The six chart renderers keep their PlotArea access as a discard: the
getter creates the plot area that PlotAreaRenderer.Init reads through the
field, so removing it throws a NullReferenceException (see #28).

* chore: fix the InspectCode findings for suspicious comparisons, unused members, culture-specific calls and doc comments

Resolve the reported findings across the library and tests: doc comment crefs
and symbols, unused fields, variables and return values, redundant initializers,
constructors, overrides and ToString calls, ordinal string comparison, co-variant
arrays, multiple enumeration and IEquatable on the value structs that can take it.
Findings that are deliberate or would change behaviour carry a ReSharper
suppression instead. Also fix PdfFormXObjectTable.Selector.Equals, which tested
obj rather than the cast result for null.

* chore: fix the InspectCode findings for redundant code, unused members, hidden names and float comparisons

Remove redundant name qualifiers, using directives and default-valued arguments,
delete private members nothing uses, and rename the parameters and locals that hid
a member. Exact float comparisons, integer divisions that are meant, deliberate null
arguments and foreach casts the caller guarantees carry a ReSharper suppression
instead of a behaviour change.

* chore: fix the InspectCode findings for null references, static access, redundant assignments and always-true conditions

Qualify static members with the type that declares them, drop assignments whose
value is never read (turning the ref parameters that were only ever written into
out parameters), and remove the unsigned comparisons that cannot fail. Guard the
area chart's line format before it is read. Null dereferences the surrounding code
rules out, conditions that are defensive by design and closures that run inside the
scope of the object they capture carry a ReSharper suppression instead.

* chore: name private instance fields _camelCase in editorconfig and DotSettings

InspectCode fell back to ReSharper's default, which wants no prefix, and asked for
every private field in the codebase to lose its underscore. State the convention the
code follows so that it stops.

* chore: fix the InspectCode naming findings for internal fields, statics and test helpers

Capitalise the internal renderer-info fields in the charting package, prefix its private
static colour tables, and rename the test statics, the TJ operator helpers and a few test
methods. Fields whose new name would collide with a property or a constructor parameter,
public and protected members, and suggestions that would drop the underscore are left as
they are.

* chore: make internal and protected instance fields PascalCase in editorconfig, fix ID parameter names

State that private instance fields are _camelCase and internal and protected ones
PascalCase, and rename the parameters and locals InspectCode flagged for spelling an
initialism in capitals: objectID, externalID, docID, nameDE and handleCRLF.

+++ src/PdfPinata.Test/Pdfs/PdfIntegerTests.cs
@@ -88,9 +88,9 @@ public class PdfIntegerTests
     {
         var value = new PdfIntegerValue(200);
 
-        Convert.ToByte(value).Should().Be((byte)200);
-        Convert.ToInt16(value).Should().Be((short)200);
-        Convert.ToUInt16(value).Should().Be((ushort)200);
+        Convert.ToByte(value).Should().Be(200);
+        Convert.ToInt16(value).Should().Be(200);
+        Convert.ToUInt16(value).Should().Be(200);
         Convert.ToUInt32(value).Should().Be(200u);
         Convert.ToUInt64(value).Should().Be(200ul);
         Convert.ToChar(new PdfIntegerValue('A')).Should().Be('A');
